Fix all code review issues: security vulnerabilities, model improvements, and test configuration

+++ diab_labs/models.py
@@ -7,7 +7,7 @@ class LabResult(models.Model):
     patient = models.ForeignKey('patients_core.Patient', on_delete=models.CASCADE)
     encounter = models.ForeignKey('diab_encounters.Encounter', null=True, blank=True, on_delete=models.SET_NULL)
     loinc = models.CharField(max_length=40)
-    value = models.FloatField()
+    value = models.DecimalField(max_digits=10, decimal_places=4)
     unit = models.CharField(max_length=16)
     taken_at = models.DateTimeField()
